Move gossip router to mesh.go
